+++ tests/dir_cases/test1-mojo-expected/bar.mojo
@@ -1,2 +1,2 @@
-fn bar1() -> Int64:
+fn bar1() -> Int:
     return 0
